_pack/requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
python-dotenv>=1.0
requests>=2.31
# Prophet is optional; install as needed (can be heavy)
prophet>=1.1 ; platform_system != 'Windows'

streamlit>=1.36
